made the code neat
